Made the output spiffier, and added a spinner
The only thing left to do that I really have any motivation to undertake
is adding in the language transition labels.  Right now, this would only
be a matter of creating a mapping from language keys to their english
names, and then appending that transition text to their arrow's div.

But it's really just a small athletic thing, so It's not a big priority